app/requirements.txt
apscheduler==3.6.3; python_version >= "3.7"
asgiref==3.5.2; python_version >= "3.8"
cachetools==4.2.2; python_version >= "3.7" and python_version < "4.0"
certifi==2022.6.15; python_version >= "3.7"
django-dotenv==1.4.2
django==4.1; python_version >= "3.8"
humanize==4.3.0; python_version >= "3.7"
psycopg2-binary==2.9.3; python_version >= "3.6"
python-dateutil==2.8.2; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.3.0")
python-telegram-bot==13.13; python_version >= "3.7"
pytz-deprecation-shim==0.1.0.post0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
pytz==2022.2.1; python_version >= "3.7"
six==1.16.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.7"
sqlparse==0.4.2; python_version >= "3.8"
tornado==6.2; python_version >= "3.7"
tzdata==2022.2; sys_platform == "win32" and python_version >= "3.8" and platform_system == "Windows" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7")
tzlocal==4.2; python_version >= "3.7"
